tüm sayfalara yetki kontrol eklendi

+++ Wms12m.Presentation/Controllers/StorageController.cs
@@ -16,6 +16,7 @@ namespace Wms12m.Presentation.Controllers
         /// </summary>
         public ActionResult Index()
         {
+            if (CheckPerm("Storage", PermTypes.Reading) == false) return Redirect("/");
             using (KabloEntities dbx = new KabloEntities())
             {
                 ViewBag.KabloDepoID = new SelectList(dbx.depoes.ToList(), "id", "depo1");
@@ -25,8 +26,9 @@ namespace Wms12m.Presentation.Controllers
         /// <summary>
         /// listesi
         /// </summary>
-        public ActionResult StoreGridPartial(string Id)
+        public PartialViewResult StoreGridPartial(string Id)
         {
+            if (CheckPerm("Storage", PermTypes.Reading) == false) return null;
             List<Depo> _List = new List<Depo>();
             _List = Id == "Locked" ? Store.GetList().Where(a => a.Aktif == true).ToList() : Id == "noLocked" ? Store.GetList().Where(a => a.Aktif == false).ToList() : Store.GetList();
             return PartialView("_StoreGridPartial", _List);
@@ -34,8 +36,9 @@ namespace Wms12m.Presentation.Controllers
         /// <summary>
         /// düzenle
         /// </summary>
-        public ActionResult StoreDetailPartial(string Id)
+        public PartialViewResult StoreDetailPartial(string Id)
         {
+            if (CheckPerm("Storage", PermTypes.Reading) == false) return null;
             var item = Convert.ToInt16(Id == "" ? "0" : Id) > 0 ? Store.Detail(Convert.ToInt16(Id)) : new Depo() { Aktif = false };
             using (KabloEntities dbx = new KabloEntities())
             {
@@ -48,6 +51,7 @@ namespace Wms12m.Presentation.Controllers
         /// </summary>
         public JsonResult Delete(string Id)
         {
+            if (CheckPerm("Storage", PermTypes.Deleting) == false) return Json(new Result(false, "Yetkiniz yok"), JsonRequestBehavior.AllowGet);
             Result _Result = Store.Delete(string.IsNullOrEmpty(Id) ? 0 : Convert.ToInt32(Id));
             return Json(_Result, JsonRequestBehavior.AllowGet);
             
@@ -55,8 +59,9 @@ namespace Wms12m.Presentation.Controllers
         /// <summary>
         /// kayıt işlemleri
         /// </summary>
-        public ActionResult StoreOperation(Depo P)
+        public JsonResult StoreOperation(Depo P)
         {
+            if (CheckPerm("Storage", PermTypes.Writing) == false) return Json(new Result(false, "Yetkiniz yok"), JsonRequestBehavior.AllowGet);
             Result _Result = Store.Operation(P);
             return Json(_Result, JsonRequestBehavior.AllowGet);
         }
